Do not notify about bwauths out of sync
Stop sending the message "NOTICE: Bandwidth authorities have a
substantially different number of measured entries" when there are
bandwidth authorities running sbws, since it is known that sbws
is reporting less number of relays in its data.

This is a temporal hack that can be reverted if/when we change
sbws to include more relays.